CRUD Operation Using React, Nodejs, Express, MongoDb
React is a front-end library developed by Facebook. It's used for handling front end for web and mobile apps. ReactJS allows us to create reusable UI components.
Node is an open source server framework.It is used to develop I/O intensive web applications like video streaming sites, single-page applications, and other web applications.
Express is a web framework for Node.js. It is a fast, robust and asynchronous in nature.
MongoDB is a No SQL database. It is an open-source, cross-platform, document-oriented database written in C++.